+++ src/dashboardWebView/components/Contents/Item.tsx
@@ -8,6 +8,7 @@ import { DateField } from '../DateField';
 import { Status } from '../Status';
 import { Messenger } from '@estruyf/vscode/dist/client';
 import { DashboardViewType } from '../../models';
+import { ContentActions } from './ContentActions';
 
 export interface IItemProps extends Page {}
 
@@ -35,6 +36,7 @@ export const Item: React.FunctionComponent<IItemProps> = ({ fmFilePath, date, ti
       <li className="relative">
         <button className={`group cursor-pointer flex flex-wrap items-start content-start h-full w-full bg-gray-50 dark:bg-vulcan-200 text-vulcan-500 dark:text-whisper-500 text-left overflow-hidden shadow-md dark:shadow-none hover:shadow-xl dark:hover:bg-vulcan-100 border border-gray-200 dark:border-vulcan-50`}
                 onClick={openFile}>
+
           <div className="relative h-36 w-full overflow-hidden border-b border-gray-100 dark:border-vulcan-100 dark:group-hover:border-vulcan-200">
             {
               pageData[PREVIEW_IMAGE_FIELD] ? (
@@ -46,6 +48,12 @@ export const Item: React.FunctionComponent<IItemProps> = ({ fmFilePath, date, ti
               )
             }
           </div>
+
+          <ContentActions 
+            title={title}
+            path={fmFilePath}
+            scripts={settings?.scripts}
+            onOpen={openFile} />
 
           <div className="p-4 w-full">
             <div className={`flex justify-between items-center`}>