+++ src/dashboardWebView/components/Menu/index.ts
@@ -1,3 +1,5 @@
+export * from './ActionMenuButton';
 export * from './MenuButton';
 export * from './MenuItem';
 export * from './MenuItems';
+export * from './QuickAction';